+++ src/react-query-demo/features/all-entries-meta/index.js
@@ -1,0 +1,1 @@
+export { AllEntriesMeta } from "./all-entries-meta";
