+++ src/react-query-demo/features/child-entries-meta/child-entries-meta.js
@@ -1,28 +1,29 @@
 import React, { useState } from "react";
-import { useGetEntryBody } from "./hooks";
-import { EntrySelector } from "./entry-selector";
-import { PrintJson } from "./components";
+import { useGetChildEntriesMeta } from "../../hooks";
+import { EntrySelector } from "../entry-selector";
+import { PrintJson } from "../../components";
 
-export function EntryBody() {
+export function ChildEntriesMeta() {
   const [selectedEntry, setSelectedEntry] = useState("");
+
   const { isLoading, isFetching, isSuccess, isError, data } =
-    useGetEntryBody(selectedEntry);
+    useGetChildEntriesMeta(selectedEntry);
 
   const handleOnChange = (event) => setSelectedEntry(event.target.value);
 
   return (
     <div>
-      <h3>Entry Body</h3>
+      <h3>Child Entries Meta</h3>
       <p>
-        <i>useGetEntryBody(entryId)</i>
+        <i>useGetChildEntries(entryId)</i>
       </p>
       <EntrySelector onChange={handleOnChange} value={selectedEntry} />
       {selectedEntry && (isLoading || isFetching) ? (
-        <div>Loading entry body</div>
+        <div>Loading Child Entries Meta</div>
       ) : (
         <>
           {isSuccess && <PrintJson data={data} />}
-          {isError && <div>Error fetching entry body</div>}
+          {isError && <div>Error fetching Child Entries Meta</div>}
         </>
       )}
       <hr />
