+++ src/react-query-demo/features/collection-selector/collection-selector.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useGetCollections } from "./hooks";
+import { useGetCollections } from "../../hooks";
 
 export function CollectionSelector({ onChange, value = "", ...otherProps }) {
   const { isLoading, isError, isFetching, data } = useGetCollections();
