+++ src/react-query-demo/features/collection-selector/index.js
@@ -1,0 +1,1 @@
+export { CollectionSelector } from "./collection-selector";
