+++ src/react-query-demo/features/entries-by-collection/index.js
@@ -1,0 +1,1 @@
+export { EntriesByCollection } from "./entries-by-collection";
