+++ src/react-query-demo/features/entry-body/index.js
@@ -1,0 +1,1 @@
+export { EntryBody } from "./entry-body";
