+++ src/react-query-demo/features/entry-meta/index.js
@@ -1,0 +1,1 @@
+export { EntryMeta } from "./entry-meta";
