+++ src/react-query-demo/features/entry-selector/index.js
@@ -1,0 +1,1 @@
+export { EntrySelector } from "./entry-selector";
