+++ src/react-query-demo/index.js
@@ -1,14 +1,6 @@
 import React from "react";
 import { QueryClient, QueryClientProvider } from "react-query";
-import { Container } from "./components";
-import { AllEntriesMeta } from "./all-entries-meta";
-import { EntriesByCollection } from "./entries-by-collection";
-import { EntryMeta } from "./entry-meta";
-import { EntryBody } from "./entry-body";
-import { ChildEntriesMeta } from "./child-entries-meta";
-import { SiblingEntriesMeta } from "./sibling-entries-meta";
-import { EntryEditor } from "./entry-editor/entry-editor";
-import { EntryCreator } from "./entry-creator/entry-creator";
+import { MainView } from "./views/main-view";
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -22,24 +14,7 @@ const queryClient = new QueryClient({
 export function ReactQueryDemo() {
   return (
     <QueryClientProvider client={queryClient}>
-      <Container
-        leftSlot={
-          <React.Fragment>
-            <AllEntriesMeta />
-            <EntriesByCollection />
-            <EntryMeta />
-            <EntryBody />
-            <ChildEntriesMeta />
-            <SiblingEntriesMeta />
-          </React.Fragment>
-        }
-        rightSlot={
-          <div>
-            <EntryEditor />
-            <EntryCreator />
-          </div>
-        }
-      />
+      <MainView />
     </QueryClientProvider>
   );
 }
