+++ src/react-query-demo/views/index.js
@@ -1,0 +1,1 @@
+export { MainView } from "./main-view";
